Change to use calculated speed  on OSD.
After manually compiling mpv and debugging, I've found that the speed set with set_property is not truncated or rounded, only get_property does this. This means we do get very precise matching of the video to the monitor with a speed setting with more than 6 digits after the decimal.